Visualizations with Grafana and Hosting
Set up AWS S# bucket to host website on, and 2nd S3 bucket that accepted uploaded files, using API Gateway and Lambda to connect them. Visible in slides, but not in demo because of AWS error on day of demo, "Your access to the course/lab is not yet aproved. Please contact your instructor."